Extract OperationsLedgerPostingService from the continuity workflow facade
Increment 3a of the read-service/workflow monolith decomposition (Phase 3).
Splits the ledger-posting concern out of the 3,845-line
src/Meridian.FinancialOperations/OperationsContinuity/OperationsContinuityWorkflowService.cs
into a dedicated single-responsibility service, keeping the workflow service as a
behavior-preserving facade over the unchanged IOperationsContinuityWorkflowService.

New files (same namespace):
- OperationsLedgerPostingService.cs (internal sealed): owns the three optional
  ledger/security dependencies (ILedgerJournalStore, IOperationsContinuityTransactionalCommitStore,
  ISecurityMasterQueryService) and the ledger cluster moved verbatim — ValidateLedgerPostRequest,
  TryBuildJournalWrite, ValidateJournalCandidate, ResolveAuthoritativeSecurityStatusesAsync, and
  the pure journal/dimension/lineage helpers. Adds a thin delegation seam: IsLedgerStoreAvailable
  and CommitOrAppendAsync(...) returning a LedgerCommitOutcome (Committed | Rejected | Proceed)
  that encapsulates the transactional-commit vs journal-append-then-fall-through store branch,
  preserving the exact short-circuit ordering, the LedgerValidationException -> blocker mapping
  (byte-identical code/message/ex.Message), and the graceful null-store path.
- OperationsContinuityWorkflowText.cs (internal static): the shared redaction regexes,
  RedactSensitiveText, and NormalizeEvidence — the members used by both the ledger and
  projection concerns — so the sub-service references one copy with no back-dependency on
  the facade.

Facade changes: the 6-arg constructor is byte-identical, so both DI registration sites and
the 88-test oracle (OperationsContinuityWorkflowServiceTests) are untouched. The constructor
now builds _ledgerPosting internally; PostLedgerEntriesAsync delegates
validate/resolve/build/commit to it while keeping repository, audit-store, status-derivation,
CloneWorkflow, ToDtoAsync, and the Success/Failure envelope in the facade. The
LEDGER_JOURNAL_STORE_UNAVAILABLE blocker stays facade-owned in both request and service blocker
lists. BuildLedgerDraftAsync/ValidateLedgerDraftAsync are unchanged.

Core OperationsContinuityWorkflowService.cs drops 3,845 -> 3,142 lines; ratchet baseline
retightened.
